# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Start API",
        "description": "AgroBozor uchun backend API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://daladan-api.loc/api/v1",
            "description": "Local dev server (OSPanel)"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "php artisan serve"
        }
    ],
    "paths": {
        "/profile/ads": {
            "get": {
                "tags": [
                    "Profile",
                    "Ads"
                ],
                "summary": "O'z reklamalari ro'yxati",
                "description": "index() — GET /profile/ads",
                "operationId": "7a9731f2689f980407da9d1da8ae61b5",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "E'lonlar ro'yxati"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile",
                    "Ads"
                ],
                "summary": "Yangi e'lon yaratish",
                "description": "store() — POST /profile/ads",
                "operationId": "e5d7980dbd0c715513ff692d8cf3ec26",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "subcategory_id",
                                    "title"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "subcategory_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "district": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "unit": {
                                        "type": "string",
                                        "enum": [
                                            "kg",
                                            "ton",
                                            "bag",
                                            "box",
                                            "piece"
                                        ],
                                        "nullable": true
                                    },
                                    "media[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "E'lon yaratildi"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/ads/{ad}": {
            "get": {
                "tags": [
                    "Profile",
                    "Ads"
                ],
                "summary": "Bitta o'z reklamasini ko'rish",
                "description": "show() — GET /profile/ads/{ad}",
                "operationId": "37fefae3414d73b9996bf7a4b362b78e",
                "parameters": [
                    {
                        "name": "ad",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "E'lon"
                    },
                    "404": {
                        "description": "Topilmadi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile",
                    "Ads"
                ],
                "summary": "PUT alias (JSON body)",
                "operationId": "98538465f42c5817e7a729fa6ccc3b33",
                "parameters": [
                    {
                        "name": "ad",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yangilandi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile",
                    "Ads"
                ],
                "summary": "E'lonni yangilash (form-data/media uchun POST, JSON uchun PUT/PATCH)",
                "description": "update() — POST /profile/ads/{ad}  (yoki PUT / PATCH)",
                "operationId": "d8ced18cdf278e093c9814ac37544c50",
                "parameters": [
                    {
                        "name": "ad",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "subcategory_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "district": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "unit": {
                                        "type": "string",
                                        "enum": [
                                            "kg",
                                            "ton",
                                            "bag",
                                            "box",
                                            "piece"
                                        ],
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "sold",
                                            "deleted"
                                        ]
                                    },
                                    "delete_media_ids[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "media[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "E'lon yangilandi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile",
                    "Ads"
                ],
                "summary": "O'z reklamasini o'chirish",
                "description": "destroy() — DELETE /profile/ads/{ad}",
                "operationId": "102a0930d0df2b4f5c7409f368aa9fe3",
                "parameters": [
                    {
                        "name": "ad",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "O'chirildi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profile",
                    "Ads"
                ],
                "summary": "PATCH alias (JSON body)",
                "operationId": "eb487286a597fd7f919040798130f1fd",
                "parameters": [
                    {
                        "name": "ad",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yangilandi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}